// packages/styles/src/styles.d.ts
export declare const toRem: {
    (value: number): string;
    displayName: string;
    __docgenInfo: {
        "description": string;
        "displayName": string;
        "props": {
            "toString": {
                "defaultValue": any;
                "description": string;
                "name": string;
                "required": boolean;
                "type": {
                    "name": string;
                };
            };
            "toFixed": {
                "defaultValue": any;
                "description": string;
                "name": string;
                "required": boolean;
                "type": {
                    "name": string;
                };
            };
            "toExponential": {
                "defaultValue": any;
                "description": string;
                "name": string;
                "required": boolean;
                "type": {
                    "name": string;
                };
            };
            "toPrecision": {
                "defaultValue": any;
                "description": string;
                "name": string;
                "required": boolean;
                "type": {
                    "name": string;
                };
            };
            "valueOf": {
                "defaultValue": any;
                "description": string;
                "name": string;
                "required": boolean;
                "type": {
                    "name": string;
                };
            };
            "toLocaleString": {
                "defaultValue": any;
                "description": string;
                "name": string;
                "required": boolean;
                "type": {
                    "name": string;
                };
            };
        };
    };
};
export declare const setAlpha: (hex: string, alpha: number) => string;
export declare const STORAGE_THEME_ID: string;
export declare const STORAGE_DEFAULT_FILTER_ID: string;
export declare const STORAGE_NOTIFICATIONS_ID: string;
export declare const STORAGE_CUSTOM_COLORS: string;
export declare const STORAGE_THEME_OVERRIDE: string;
export declare const THEME_DARK = "dark";
export declare const THEME_LIGHT = "light";
export declare const THEME_CUSTOM = "custom";
export declare const catastropheeColors: {
    lime: string;
    green: string;
    cyan: string;
    blue: string;
    royalBlue: string;
    purple: string;
    violet: string;
    pink: string;
    red: string;
    orange: string;
    pastel: string;
    yellow: string;
    darkGrey: string;
    grey: string;
};
// ...
export declare const Borders: {
    radius: string;
};
export declare const Opacity: {
    light: number;
};
export declare const Margins: {
    default: string;
    relaxed: string;
    regular: string;
    slim: string;
};
export declare const PaddingsNumbers: {
    default: number;
    relaxed: number;
    regular: number;
    slim: number;
};
export declare const Paddings: {
    default: string;
    relaxed: string;
    regular: string;
    slim: string;
};
export declare const Size: {
    appBarHeight: number;
    appMargin: number;
    border: number;
    iconSize: number;
    iconSpacing: number;
    itemSpacing: number;
    keyLine: number;
};
export declare const superExtraLight = 100;
export declare const extraLight = 200;
export declare const Light = 300;
export declare const Regular = 400;
export declare const Medium = 500;
export declare const Bold = 700;
export declare const Black = 900;
export declare const Family = "\"Circular EP\", -apple-system, \"Helvetica Neue\", \"Lucida Grande\", sans-serif";
// ...
export declare const Color: {
    toggleBackground: string;
    toggleThumb: string;
    completed: string;
    decoration: string;
    done: string;
    inProgress: string;
    onCompleted: string;
    onDone: string;
    onInProgress: string;
    onOverdue: string;
    onSuccess: string;
    onPending: string;
    overdue: string;
    pending: string;
    rejection: string;
    success: string;
    shadow: string;
    warning: string;
    brand: string;
    onBrand: string;
    highlight: string;
    primary: string;
    primary10: string;
    primary100: string;
    primary200: string;
    primary300: string;
    primary400: string;
    onPrimary: string;
    onPrimary10: string;
    onPrimary200: string;
    onPrimary100: string;
    onPrimary300: string;
    onPrimary400: string;
    secondary: string;
    onSecondary: string;
    background: string;
    surface: string;
    surface10: string;
    surface100: string;
    error: string;
    onBackground: string;
    onBackgroundVariant: string;
    onSurface: string;
    onError: string;
    dark: string;
    light: string;
    foreground: string;
};